#79 - refactor: 엔티티 클래스 패키지 이동
