+++ src/main/java/com/gxdxx/instagram/config/AuditorAwareImpl.java
@@ -1,6 +1,5 @@
 package com.gxdxx.instagram.config;
 
-import com.gxdxx.instagram.entity.User;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
