+++ src/main/java/com/gxdxx/instagram/config/jwt/TokenProvider.java
@@ -1,6 +1,6 @@
 package com.gxdxx.instagram.config.jwt;
 
-import com.gxdxx.instagram.entity.User;
+import com.gxdxx.instagram.domain.user.domain.User;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwts;
