+++ src/main/java/com/gxdxx/instagram/domain/follow/domain/Follow.java
@@ -1,5 +1,7 @@
-package com.gxdxx.instagram.entity;
+package com.gxdxx.instagram.domain.follow.domain;
 
+import com.gxdxx.instagram.domain.user.domain.User;
+import com.gxdxx.instagram.domain.BaseEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
