+++ src/main/java/com/gxdxx/instagram/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.gxdxx.instagram.repository;
 
-import com.gxdxx.instagram.entity.Comment;
+import com.gxdxx.instagram.domain.comment.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
