+++ src/main/java/com/gxdxx/instagram/repository/ReplyRepository.java
@@ -1,6 +1,6 @@
 package com.gxdxx.instagram.repository;
 
-import com.gxdxx.instagram.entity.Reply;
+import com.gxdxx.instagram.domain.reply.domain.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReplyRepository extends JpaRepository<Reply, Long> {
