+++ src/main/java/com/gxdxx/instagram/service/comment/CommentUpdateService.java
@@ -2,7 +2,7 @@ package com.gxdxx.instagram.service.comment;
 
 import com.gxdxx.instagram.dto.request.CommentUpdateRequest;
 import com.gxdxx.instagram.dto.response.CommentUpdateResponse;
-import com.gxdxx.instagram.entity.Comment;
+import com.gxdxx.instagram.domain.comment.domain.Comment;
 import com.gxdxx.instagram.exception.CommentNotFoundException;
 import com.gxdxx.instagram.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.repository.CommentRepository;
