+++ src/main/java/com/gxdxx/instagram/service/message/MessageQueryService.java
@@ -3,8 +3,8 @@ package com.gxdxx.instagram.service.message;
 import com.gxdxx.instagram.dto.request.MessageListRequest;
 import com.gxdxx.instagram.dto.response.MessageListResponse;
 import com.gxdxx.instagram.dto.response.MessageResponse;
-import com.gxdxx.instagram.entity.ChatRoom;
-import com.gxdxx.instagram.entity.User;
+import com.gxdxx.instagram.domain.chatroom.domain.ChatRoom;
+import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.ChatRoomNotFoundException;
 import com.gxdxx.instagram.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
