+++ src/main/java/com/gxdxx/instagram/service/post/PostFeedQueryService.java
@@ -3,7 +3,7 @@ package com.gxdxx.instagram.service.post;
 import com.gxdxx.instagram.dto.request.PostFeedRequest;
 import com.gxdxx.instagram.dto.response.FeedResponse;
 import com.gxdxx.instagram.dto.response.PostFeedResponse;
-import com.gxdxx.instagram.entity.User;
+import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.UserNotFoundException;
 import com.gxdxx.instagram.repository.PostRepository;
 import com.gxdxx.instagram.repository.UserRepository;
