+++ src/main/java/com/gxdxx/instagram/service/post/PostUpdateService.java
@@ -2,7 +2,7 @@ package com.gxdxx.instagram.service.post;
 
 import com.gxdxx.instagram.dto.request.PostUpdateRequest;
 import com.gxdxx.instagram.dto.response.PostUpdateResponse;
-import com.gxdxx.instagram.entity.Post;
+import com.gxdxx.instagram.domain.post.domain.Post;
 import com.gxdxx.instagram.exception.PostNotFoundException;
 import com.gxdxx.instagram.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.repository.PostRepository;
