+++ src/main/java/com/gxdxx/instagram/service/reply/ReplyUpdateService.java
@@ -2,7 +2,7 @@ package com.gxdxx.instagram.service.reply;
 
 import com.gxdxx.instagram.dto.request.ReplyUpdateRequest;
 import com.gxdxx.instagram.dto.response.ReplyUpdateResponse;
-import com.gxdxx.instagram.entity.Reply;
+import com.gxdxx.instagram.domain.reply.domain.Reply;
 import com.gxdxx.instagram.exception.ReplyNotFoundException;
 import com.gxdxx.instagram.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.repository.ReplyRepository;
