+++ src/main/java/com/gxdxx/instagram/service/token/AccessTokenCreateService.java
@@ -3,7 +3,7 @@ package com.gxdxx.instagram.service.token;
 import com.gxdxx.instagram.config.jwt.TokenProvider;
 import com.gxdxx.instagram.dto.request.AccessTokenCreateRequest;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
-import com.gxdxx.instagram.entity.User;
+import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.RefreshTokenInvalidException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
 import com.gxdxx.instagram.repository.UserRepository;
