+++ src/main/java/com/gxdxx/instagram/service/user/UserProfileQueryService.java
@@ -1,7 +1,7 @@
 package com.gxdxx.instagram.service.user;
 
 import com.gxdxx.instagram.dto.response.UserProfileResponse;
-import com.gxdxx.instagram.entity.User;
+import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.UserNotFoundException;
 import com.gxdxx.instagram.repository.FollowRepository;
 import com.gxdxx.instagram.repository.UserRepository;
