+++ src/test/java/com/gxdxx/instagram/service/follow/FollowDeleteServiceTest.java
@@ -1,12 +1,9 @@
 package com.gxdxx.instagram.service.follow;
 
-import com.gxdxx.instagram.dto.request.FollowCreateRequest;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
-import com.gxdxx.instagram.entity.Follow;
-import com.gxdxx.instagram.entity.User;
-import com.gxdxx.instagram.exception.FollowAlreadyExistsException;
+import com.gxdxx.instagram.domain.follow.domain.Follow;
+import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.FollowNotFountException;
-import com.gxdxx.instagram.exception.InvalidRequestException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
 import com.gxdxx.instagram.repository.FollowRepository;
 import com.gxdxx.instagram.repository.UserRepository;
