+++ src/test/java/com/gxdxx/instagram/service/user/UserDeleteServiceTest.java
@@ -1,9 +1,8 @@
 package com.gxdxx.instagram.service.user;
 
 import com.gxdxx.instagram.dto.response.SuccessResponse;
-import com.gxdxx.instagram.entity.User;
+import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.PasswordNotMatchException;
-import com.gxdxx.instagram.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
 import com.gxdxx.instagram.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
